requirements.txt
python-frontmatter
textstat
nltk
